ibmcloud: add dedicated host group configuration
Add support for dedicated host groups by allowing
users to specify a host group ID.

If a host group ID is provided, the pod will be
created on the specified dedicated host group.

If both a host group ID and a host ID are provided,
the host ID will take precedence.
When creation fails using the host ID, the provider
will attempt to create the instance using the host group ID.

Signed-off-by: Patrik Fodor <patrik.fodor@ibm.com>